упрощен круд меню и подменю

+++ app/api/menus/crud.py
@@ -3,14 +3,13 @@ from sqlalchemy.orm.exc import FlushError, NoResultFound
 
 from app.database.models import Menu
 from app.database.schemas import MenuPost
-from app.database.services import check_objects, check_unique_menu
+from app.database.services import check_unique_menu
 
 
 def get_menu_by_id(db: Session, id: str):
     """Получение меню по id."""
-    try:
-        check_objects(db=db, menu_id=id)
-    except NoResultFound:
+    current_menu = db.query(Menu).filter(Menu.id == id).first()
+    if not current_menu:
         raise NoResultFound("menu not found")
     return db.query(Menu).filter(Menu.id == id).first()
 
@@ -38,15 +37,13 @@ def create_menu(db: Session, menu: MenuPost):
 
 def update_menu(db: Session, menu_id: str, updated_menu: MenuPost):
     """Изменение меню по id."""
-    try:
-        check_objects(db=db, menu_id=menu_id)
-    except NoResultFound:
+    current_menu = get_menu_by_id(db=db, id=menu_id)
+    if not current_menu:
         raise NoResultFound("menu not found")
     try:
         check_unique_menu(db=db, menu=updated_menu)
     except FlushError:
         raise FlushError("Меню с таким названием уже есть")
-    current_menu = get_menu_by_id(db=db, id=menu_id)
     current_menu.title = updated_menu.title
     current_menu.description = updated_menu.description
     db.merge(current_menu)
@@ -57,10 +54,8 @@ def update_menu(db: Session, menu_id: str, updated_menu: MenuPost):
 
 def delete_menu(db: Session, menu_id: str):
     """Удаление меню по id."""
-    try:
-        check_objects(db=db, menu_id=menu_id)
-    except NoResultFound:
-        raise NoResultFound("menu not found")
     current_menu = get_menu_by_id(db=db, id=menu_id)
+    if not current_menu:
+        raise NoResultFound("menu not found")
     db.delete(current_menu)
     db.commit()
